Handle pause/exit commands
Reset the end attribute to 0 during the release process.
This ensures that once a game session ends, users can start new session.
We can run xo-user again without any residual state from previous one.
Resolves the issue of garbled output.
Because the data we read from the device is not a string.
Add the terminator to tell printf where to stop printing.