Implement the board record
This commit enables the kernel module to record previous game boards.
It can store 15 boards in total.
It will be displayed when the user quits this game.
Use a unsigned 64-bit table to record previous game boards.
Use `ioctl` to send table and the total moves in each game.
The table and moves mentioned above is from kernel mode to user level.
Finally, we decodes every move in user space, and print in terminal.